Fix : Change id datetype

+++ schedule.sql
@@ -5,7 +5,7 @@ USE schedule_db;
 DROP TABLE IF EXISTS schedule_tbl, user_tbl;
 
 CREATE TABLE user_tbl (
-    id INT AUTO_INCREMENT NOT NULL PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT NOT NULL PRIMARY KEY,
     name VARCHAR(20) NOT NULL,
     email VARCHAR(30) NOT NULL,
     password VARCHAR(255) NOT NULL,
@@ -14,8 +14,8 @@ CREATE TABLE user_tbl (
 ) ENGINE = InnoDB;
 
 CREATE TABLE schedule_tbl (
-    id INT AUTO_INCREMENT NOT NULL PRIMARY KEY,
-    user_id INT NOT NULL,
+    id BIGINT AUTO_INCREMENT NOT NULL PRIMARY KEY,
+    user_id BIGINT NOT NULL,
     title VARCHAR(30) NOT NULL,
     content TINYTEXT NOT NULL,
     created_at DATETIME NOT NULL,
